BUG: Fix `nanpercentile` ignoring the dtype of all-nan arrays

+++ numpy/lib/nanfunctions.py
@@ -1421,7 +1421,8 @@ def _nanquantile_1d(arr1d, q, overwrite_input=False, interpolation='linear'):
     arr1d, overwrite_input = _remove_nan_1d(arr1d,
         overwrite_input=overwrite_input)
     if arr1d.size == 0:
-        return np.full(q.shape, np.nan)[()]  # convert to scalar
+        # convert to scalar
+        return np.full(q.shape, np.nan, dtype=arr1d.dtype)[()]
 
     return function_base._quantile_unchecked(
         arr1d, q, overwrite_input=overwrite_input, interpolation=interpolation)
